requirements.txt
pandas>=2.2.0
python-dotenv>=1.0.1
faiss-cpu>=1.8.0.post1
numpy>=1.26.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
streamlit>=1.44.0
